backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# OCR
easyocr>=1.7.1
Pillow>=10.0.0
opencv-python-headless>=4.9.0.80

# AI/LLM
groq>=0.4.1

# Utils
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
